% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{build_wiki_url}
\alias{build_wiki_url}
\title{Build Wiki URL}
\usage{
build_wiki_url(wiki, type = NULL, page = NULL, format = "json",
  action = "parse", redirects = TRUE, api = FALSE)
}
\description{
Supports both wiki page urls and their equivalent API calls.
}
\examples{
build_wiki_url("en", "wikipedia", "Malus domestica")
build_wiki_url("en", "wikipedia", "Malus domestica", api = TRUE)
build_wiki_url(parse_wiki_url("https://en.wikipedia.org/wiki/Malus_domestica"))
}
\seealso{
Other Wiki functions: \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_page}},
  \code{\link{parse_wikispecies_page}}
}

