% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipages.R
\name{get_wiki_page}
\alias{get_wiki_page}
\title{Get MediaWiki Page from API}
\usage{
get_wiki_page(url, ...)
}
\arguments{
\item{url}{() MediaWiki page url.}

\item{...}{Arguments passed to \code{\link{build_wiki_url}} if \code{url} is a static page url.}
}
\description{
Supports both static page urls and their equivalent API calls.
}
\examples{
str(get_wiki_page("https://en.wikipedia.org/wiki/Malus_domestica"))
}
\seealso{
Other MediaWiki functions: \code{\link{build_wiki_url}},
  \code{\link{parse_wiki_page}},
  \code{\link{parse_wiki_url}}
}

