% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wiki_page}
\alias{parse_wiki_page}
\title{Parse Wiki Page}
\usage{
parse_wiki_page(page, types = c("langlinks", "iwlinks", "externallinks"))
}
\description{
Parses common properties from Wikimedia API page results.
}
\details{
NOTE: Currently not parsed:
title, displaytitle, pageid, revid, redirects[], text[1], categories[], links[], templates[], images[], sections[], properties[]
}
\examples{
pg <- get_wiki_page("https://en.wikipedia.org/wiki/Malus_domestica")
str(parse_wiki_page(pg))
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_page}},
  \code{\link{parse_wikispecies_page}}
}

