% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wiki_url}
\alias{parse_wiki_url}
\title{Parse Wiki URL}
\usage{
parse_wiki_url(url, api = FALSE)
}
\description{
Supports both wiki page urls and their equivalent API calls.
}
\examples{
parse_wiki_url("https://en.wikipedia.org/wiki/Malus_domestica")
parse_wiki_url("https://en.wikipedia.org/w/api.php?page=Malus_domestica")
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_page}},
  \code{\link{parse_wikicommons_page}},
  \code{\link{parse_wikispecies_page}}
}

