% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipages.R
\name{parse_wikicommons_page}
\alias{parse_wikicommons_page}
\title{Parse Wikimedia Commons Page}
\usage{
parse_wikicommons_page(page, types = c("langlinks", "iwlinks",
  "externallinks", "common_names"))
}
\arguments{
\item{page}{(\code{\link[httr]{response}}) Result of \code{\link{get_wiki_page}}.}

\item{types}{(character) List of properties to parse.}
}
\description{
Parses properties from the result of a Wikimedia Commons API page call.
}
\examples{
pg <- get_wiki_page("https://commons.wikimedia.org/wiki/Malus_domestica")
str(parse_wikicommons_page(pg))
pg <- get_wiki_page("https://commons.wikimedia.org/wiki/Abelmoschus")
str(parse_wikicommons_page(pg))
}
