% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wikipedia_page}
\alias{parse_wikipedia_page}
\title{Parse Wikipedia Page}
\usage{
parse_wikipedia_page(page, types = c("langlinks", "iwlinks", "externallinks",
  "common_names"))
}
\description{
Parse Wikipedia Page
}
\examples{
pg <- get_wiki_page("https://en.wikipedia.org/wiki/Malus_domestica")
str(parse_wikipedia_page(pg))
pg <- get_wiki_page("https://en.wikipedia.org/wiki/Abelmoschus")
str(parse_wikipedia_page(pg)) # no common names
}

