% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikitaxa-package.R
\docType{data}
\name{wikipedias}
\alias{wikipedias}
\title{List of Wikipedias}
\description{
data.frame of 295 rows, with 3 columns:
\itemize{
\item language - language
\item language_local - language in local name
\item wiki - langugae code for the wiki
}
}
\details{
From \url{https://meta.wikimedia.org/wiki/List_of_Wikipedias}
}
\keyword{data}
