% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikitaxa-package.R
\docType{package}
\name{wikitaxa-package}
\alias{wikitaxa}
\alias{wikitaxa-package}
\title{Taxonomic Information from Wikipedia}
\description{
Taxonomic Information from Wikipedia
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

