% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipages.R
\name{wt_wiki_page}
\alias{wt_wiki_page}
\title{Get MediaWiki Page from API}
\usage{
wt_wiki_page(url, ...)
}
\arguments{
\item{url}{(character) MediaWiki page url.}

\item{...}{Arguments passed to \code{\link[=wt_wiki_url_build]{wt_wiki_url_build()}} if \code{url}
is a static page url.}
}
\value{
an \pkg{httr} response
}
\description{
Supports both static page urls and their equivalent API calls.
}
\examples{
str(wt_wiki_page("https://en.wikipedia.org/wiki/Malus_domestica"))
}
\seealso{
Other MediaWiki functions: \code{\link{wt_wiki_page_parse}},
  \code{\link{wt_wiki_url_build}},
  \code{\link{wt_wiki_url_parse}}
}
