% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikicommons.R
\name{wt_wikicommons}
\alias{wt_wikicommons}
\alias{wt_wikicommons_parse}
\alias{wt_wikicommons_search}
\title{WikiCommons}
\usage{
wt_wikicommons(name, utf8 = TRUE)

wt_wikicommons_parse(page, types = c("langlinks", "iwlinks", "externallinks",
  "common_names", "classification"), tidy = FALSE)

wt_wikicommons_search(query, limit = 10, offset = 0, utf8 = TRUE, ...)
}
\arguments{
\item{name}{(character) Wiki name - as a page title}

\item{utf8}{(boolean) If \code{TRUE}, encodes most (but not all) non-ASCII
characters as UTF-8 instead of replacing them with hexadecimal escape
sequences.}

\item{page}{(\code{\link[httr:response]{httr::response()}}) Result of \code{\link[=wt_wiki_page]{wt_wiki_page()}}.}

\item{types}{(character) List of properties to parse}
}
\value{
\code{wt_wikicommons} returns a list, with slots:
\itemize{
\item langlinks - language page links
\item externallinks - external links
\item common_names - a data.frame with \code{name} and \code{language} columns
}

\code{wt_wikicommons_parse} returns a list
}
\description{
WikiCommons
}
\examples{
# high level
wt_wikicommons(name = "Malus domestica")
wt_wikicommons(name = "Poa annua")
wt_wikicommons(name = "Quercus")

# low level
pg <- wt_wiki_page("https://commons.wikimedia.org/wiki/Malus_domestica")
wt_wikicommons_parse(pg)
pg <- wt_wiki_page("https://commons.wikimedia.org/wiki/Abelmoschus")
wt_wikicommons_parse(pg)

# search wikicommons
wt_wikicommons_search(query = "Pinus")
wt_wikicommons_search(query = "pine tree", limit = 3)
wt_wikicommons_search(query = "pine tree", limit = 3, offset = 3)

## use search results to dig into pages
res <- wt_wikicommons_search(query = "Pinus")
lapply(res$query$search$title[1:3], wt_wikicommons)
}
