% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipages.R
\name{parse_wiki_page}
\alias{parse_wiki_page}
\title{Parse MediaWiki Page}
\usage{
parse_wiki_page(page, types = c("langlinks", "iwlinks", "externallinks"))
}
\arguments{
\item{page}{(\code{\link[httr]{response}}) Result of \code{\link{get_wiki_page}}.}

\item{types}{(character) List of properties to parse.}
}
\description{
Parses common properties from the result of a MediaWiki API page call.
}
\details{
Available properties currently not parsed:
title, displaytitle, pageid, revid, redirects[], text[1], categories[], links[], templates[], images[], sections[], properties[], ...
}
\examples{
pg <- get_wiki_page("https://en.wikipedia.org/wiki/Malus_domestica")
str(parse_wiki_page(pg))
}
\seealso{
Other MediaWiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}}, \code{\link{parse_wiki_url}}
}

