% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipages.R
\name{parse_wiki_url}
\alias{parse_wiki_url}
\title{Parse MediaWiki Page URL}
\usage{
parse_wiki_url(url)
}
\arguments{
\item{url}{(character) MediaWiki page url.}
}
\description{
Parse a MediaWiki page url into its component parts (wiki name, wiki type, and page title). Supports both static page urls and their equivalent API calls.
}
\examples{
parse_wiki_url("https://en.wikipedia.org/wiki/Malus_domestica")
parse_wiki_url("https://en.wikipedia.org/w/api.php?page=Malus_domestica")
}
\seealso{
Other MediaWiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_page}}
}

