% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikitaxa-package.R
\docType{package}
\name{wikitaxa-package}
\alias{wikitaxa-package}
\alias{wikitaxa}
\title{wikitaxa}
\description{
Taxonomic Information from Wikipedia
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Ethan Welty
}
\keyword{package}
