% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{wt_data}
\alias{wt_data}
\alias{wt_data_id}
\title{Wikidata taxonomy data}
\usage{
wt_data(x, property = NULL, ...)

wt_data_id(x, language = "en", limit = 10, ...)
}
\arguments{
\item{x}{(character) a taxonomic name}

\item{property}{(character) a property id, e.g., P486}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}

\item{language}{(character) two letter language code}

\item{limit}{(integer) records to return. Default: 10}
}
\description{
Wikidata taxonomy data
}
\examples{
\dontrun{
wt_data("Poa annua")
wt_data("Mimulus alsinoides")
wt_data("Mimulus foliatus")
wt_data("Mimulus foliatus", property = "P846")
wt_data("Mimulus foliatus", property = c("P846", "P815"))

wt_data(wt_data_id("Mimulus foliatus"))
}
}
