% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipages.R
\name{wt_wiki_page_parse}
\alias{wt_wiki_page_parse}
\title{Parse MediaWiki Page}
\usage{
wt_wiki_page_parse(page, types = c("langlinks", "iwlinks", "externallinks"))
}
\arguments{
\item{page}{(\code{\link[httr]{response}}) Result of
\code{\link{wt_wiki_page}}.}

\item{types}{(character) List of properties to parse.}
}
\value{
a list
}
\description{
Parses common properties from the result of a MediaWiki API page call.
}
\details{
Available properties currently not parsed:
title, displaytitle, pageid, revid, redirects[], text[], categories[],
links[], templates[], images[], sections[], properties[], ...
}
\examples{
pg <- wt_wiki_page("https://en.wikipedia.org/wiki/Malus_domestica")
str(wt_wiki_page_parse(pg))
}
\seealso{
Other MediaWiki functions: \code{\link{wt_wiki_page}},
  \code{\link{wt_wiki_url_build}},
  \code{\link{wt_wiki_url_parse}}
}
