% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikispecies.R
\name{wt_wikispecies}
\alias{wt_wikispecies}
\alias{wt_wikispecies_parse}
\title{WikiSpecies}
\usage{
wt_wikispecies(name, utf8 = TRUE)

wt_wikispecies_parse(page, types = c("langlinks", "iwlinks", "externallinks",
  "common_names", "classification"))
}
\arguments{
\item{name}{(character) Wiki name - as a page title}

\item{utf8}{(boolean) If \code{TRUE}, encodes most (but not all) non-ASCII
characters as UTF-8 instead of replacing them with hexadecimal escape
sequences.}

\item{page}{(\code{\link[httr]{response}}) Result of
\code{\link{wt_wiki_page}}.}

\item{types}{(character) List of properties to parse}
}
\value{
\code{wt_wikispecies} returns a list, with slots:
\itemize{
\item langlinks - language page links
\item externallinks - external links
\item common_names - a data.frame with \code{name} and \code{language} columns
}

\code{wt_wikispecies_parse} returns a list
}
\description{
WikiSpecies
}
\examples{
# high level
wt_wikispecies(name = "Malus domestica")
wt_wikispecies(name = "Poa annua")
wt_wikispecies(name = "Quercus")

# low level
pg <- wt_wiki_page("https://species.wikimedia.org/wiki/Malus_domestica")
str(wt_wikispecies_parse(pg))
pg <- wt_wiki_page("https://species.wikimedia.org/wiki/Abelmoschus")
str(wt_wikispecies_parse(pg))
}
